+++ src/index.tsx
@@ -1,5 +1,13 @@
 /* @refresh reload */
 import { render } from 'solid-js/web';
 import App from './App';
+import { AppStoreProvider } from './stores/appContext';
 
-render(() => <App />, document.getElementById('root') as HTMLElement);
+render(
+  () => (
+    <AppStoreProvider>
+      <App />
+    </AppStoreProvider>
+  ),
+  document.getElementById('root') as HTMLElement
+);
